Componentize formatTimeDifference
