+++ src/components/CommentComponent.js
@@ -5,7 +5,7 @@ import { faReply, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { updateDoc, doc, getDoc, deleteDoc } from 'firebase/firestore';
 import { db } from '../firebase';
 import fetchUserProfilePicture from '../functions/fetchUserProfilePicture';
-
+import formatTimeDifference from '../functions/formatTimeDifference';
 import { collection, getDocs, query, where } from 'firebase/firestore';
 import { auth } from '../firebase';
 import fetchUsernameWithUID from '../functions/fetchUsernameWithUID';
@@ -188,29 +188,6 @@ class CommentComponent extends Component {
     }
   };
 
-  formatTimeDifference = (timestamp) => {
-    if (isNaN(timestamp)) {
-      return 'Just Now';
-    }
-
-    const currentTime = new Date();
-    const commentTime = new Date(timestamp);
-    const differenceInSeconds = Math.floor((currentTime - commentTime) / 1000);
-
-    if (differenceInSeconds < 60) {
-      return 'just now';
-    } else if (differenceInSeconds < 3600) {
-      const minutes = Math.floor(differenceInSeconds / 60);
-      return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
-    } else if (differenceInSeconds < 86400) {
-      const hours = Math.floor(differenceInSeconds / 3600);
-      return `${hours} hour${hours > 1 ? 's' : ''} ago`;
-    } else {
-      const days = Math.floor(differenceInSeconds / 86400);
-      return `${days} day${days > 1 ? 's' : ''} ago`;
-    }
-  };
-
   render() {
     const { currentUser } = this.props;
     const { userProfilePicture, username, isReplying, replyText, replies, loadingReplies } = this.state;
@@ -227,7 +204,7 @@ class CommentComponent extends Component {
         <div className="flex-grow border p-1">
           <div className="flex justify-between">
             <Link to={`/users/${username}`} className="text-lg font-bold">{username}</Link>
-            <span className="text-sm text-gray-500">{this.formatTimeDifference(this.props.comment.createdAt)}</span>
+            <span className="text-sm text-gray-500">{formatTimeDifference(this.props.comment.createdAt)}</span>
           </div>
 
           <p className="text-sm">{this.props.comment.text}</p>
